Fix cron audit protocol and media expectations
