+++ apps/shared/OpenClawKit/Sources/OpenClawProtocol/GatewayModels.swift
@@ -5872,6 +5872,7 @@ public struct CronJob: Codable, Sendable {
     public let sessiontarget: AnyCodable
     public let wakemode: AnyCodable
     public let payload: AnyCodable
+    public let audit: [String: AnyCodable]?
     public let delivery: AnyCodable?
     public let failurealert: AnyCodable?
     public let state: [String: AnyCodable]
@@ -5890,6 +5891,7 @@ public struct CronJob: Codable, Sendable {
         sessiontarget: AnyCodable,
         wakemode: AnyCodable,
         payload: AnyCodable,
+        audit: [String: AnyCodable]?,
         delivery: AnyCodable?,
         failurealert: AnyCodable?,
         state: [String: AnyCodable])
@@ -5907,6 +5909,7 @@ public struct CronJob: Codable, Sendable {
         self.sessiontarget = sessiontarget
         self.wakemode = wakemode
         self.payload = payload
+        self.audit = audit
         self.delivery = delivery
         self.failurealert = failurealert
         self.state = state
@@ -5926,6 +5929,7 @@ public struct CronJob: Codable, Sendable {
         case sessiontarget = "sessionTarget"
         case wakemode = "wakeMode"
         case payload
+        case audit
         case delivery
         case failurealert = "failureAlert"
         case state
