+++ src/cron/types.ts
@@ -3,8 +3,7 @@ import type { FailoverReason } from "../agents/embedded-agent-helpers/types.js";
 import type { EmbeddedAgentExecutionPhase } from "../agents/embedded-agent-runner/execution-phase.js";
 import type { ChannelId } from "../channels/plugins/types.public.js";
 import type { HookExternalContentSource } from "../security/external-content.js";
-import type { CronPayloadAuditMetadata } from "./payload-audit.js";
-import type { CronJobBase } from "./types-shared.js";
+import type { CronJobBase, CronPayloadAuditMetadata } from "./types-shared.js";
 
 /** Supported schedule forms persisted in cron job specs. */
 export type CronSchedule =
